requirements.txt
rospkg
git+https://github.com/rail-berkeley/oculus_reader.git
modern_robotics
pyquaternion
scipy
funcsigs
gym
matplotlib~=3.4.2
torch~=1.8.1
torchvision~=0.9.1
pandas~=1.2.4
pillow~=8.2.0
h5py~=3.2.1
pybullet
tensorboardx~=2.2
tqdm~=4.51.0
moviepy~=1.0.3
imageio-ffmpeg
opencv-python~=4.5.2.52
scikit-image
lmdb
transformations
pyarrow

semiparametrictransfer~=0.2.dev0
setuptools~=50.3.1
numpy~=1.20.3
imageio~=2.9.0
seaborn~=0.11.1